+++ clickhouse-data/src/main/java/com/clickhouse/data/ClickHouseDataStreamFactory.java
@@ -14,7 +14,7 @@ import java.util.concurrent.TimeUnit;
 import java.util.function.Supplier;
 
 import com.clickhouse.config.ClickHouseBufferingMode;
-import com.clickhouse.data.format.ClickHouseNativeProcessor;
+import com.clickhouse.data.format.ClickHouseBinaryFormatProcessor;
 import com.clickhouse.data.format.ClickHouseRowBinaryProcessor;
 import com.clickhouse.data.format.ClickHouseTabSeparatedProcessor;
 import com.clickhouse.data.stream.BlockingPipedOutputStream;
@@ -142,8 +142,9 @@ public class ClickHouseDataStreamFactory {
                 ClickHouseFormat.RowBinaryWithDefaults == format ||
                 ClickHouseFormat.RowBinaryWithNames == format) {
             processor = new ClickHouseRowBinaryProcessor(config, input, output, columns, settings);
-        } else if (ClickHouseFormat.Native == format) {
-            processor = new ClickHouseNativeProcessor(config, input, output, columns, settings);
+        } else if (format.isBinary()) {
+            // to let outer code access input stream
+            processor = new ClickHouseBinaryFormatProcessor(config, input, output, columns, settings);
         } else if (format.isText()) {
             processor = new ClickHouseTabSeparatedProcessor(config, input, output, columns, settings);
         }